Numerous modifications to go to 1st extension of SoE
Primarily, we now have all the multiples of '2' already crossed out. We must
not use the check/set prime functions for multiples of 2, as our algorithm
that maps numbers to indices is too simple.

PRIME_COUNT calculation is simplified and is based on only one number. Several
things which were rounding improperly were fixed. The initial pCount
calculation is still incorrect, but only things which are prime are marked as
such.